Security fix to avoid deletion of item by non-admin acount.
git-svn-id: https://svn.sourceforge.jp/svnroot/nucleus-jp/nucleus-jp/trunk@712 1ca29b6e-896d-4ea0-84a5-967f57386b96

+++ euc/nucleus/libs/ITEM.php
@@ -148,7 +148,7 @@ class ITEM {
 		NucleusPlugin::_applyPluginOptions($aOptions, $itemid);
 		$manager->notify('PostPluginOptionsUpdate',array('context' => 'item', 'itemid' => $itemid, 'item' => array('title' => $i_title, 'body' => $i_body, 'more' => $i_more, 'closed' => $i_closed, 'catid' => $i_catid)));
 
-		if ($i_draftid > 0) {
+		if ($i_draftid > 0 && $member->canAlterItem($i_draftid) ) {
 			ITEM::delete($i_draftid);
 		}
 
@@ -284,9 +284,10 @@ class ITEM {
 	  * Deletes an item
 	  */
 	function delete($itemid) {
-		global $manager;
+		global $manager, $member;
 
 		$itemid = intval($itemid);
+		if (!$member->canAlterItem($itemid)) return;
 
 		$manager->notify('PreDeleteItem', array('itemid' => $itemid));
